requirements.txt
azure.cosmos
azure.communication.email
azure.identity 
python-dotenv
azure-functions
pytest
requests